requirements.txt
matplotlib==3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
protobuf<=3.20.0
pycocotools==2.0.6
pyparsing==2.4.7
requests==2.31.0
rsa==4.7
scipy==1.10.0
six==1.15.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
torch==1.13.1
torchvision==0.8.1
tqdm>=4.41.0
urllib3>=1.26.5
Werkzeug==3.0.3
zipp==3.1.0

# Logging -------------------------------------
tensorboard>=2.4.1
# wandb

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0
